refactor(sysctl): remove need for "new" data to be mutable
It was previously mutable to support macOS, which requires a mutable
pointer to the data. Upon further review, I've become convinced that
this was just oversight (people often don't think to make pointers
`const` in C) and that the "new" data is not actually modified by the
kernel. So instead, let's just cast the pointer to a mutable one.