fix: fix Clone implementations for Passwd and Group
In some cases (for example, systemd's nss module on Linux with glibc),
get{pw,gr}*_r() could fill the passwd/group structure with pointers to
static strings. In that case, the pointers don't need to be offset for
the new buffer.

Also, Group.clone() didn't properly translate the pointers in the member
list.