refactor: pass timeouts by reference
Using timeout.map_or_else(core::ptr::null, |t| t.as_ref()) is actually
UB if the timeout is an Option<TimeSpec>, since it returns a pointer to
memory that is no longer in scope. This appeared when testing in
--release mode.

This could be fixed by adding as_ref(). However, since some functions
already take an &TimeSpec, let's update the ones that don't to match for
consistency.